Fixed User Settings Update
