+++ src/components/userProfile/userSettings.js
@@ -1,84 +1,109 @@
-import React, { useContext, useState } from "react";
-import { useSelector } from "react-redux";
-import { UserSettingsForm } from "./UserSettingsForm.js";
-import { UserData } from "./userData.js";
-import { Button, Container } from "@material-ui/core";
-import { ActionsContext } from "../../context/ActionsContext";
-import { makeStyles } from "@material-ui/core/styles";
-import { useForm } from "../../hooks/useForm";
+import React, {useContext, useEffect, useState} from "react";
+import {useSelector} from "react-redux";
+import {UserSettingsForm} from "./UserSettingsForm.js";
+import {UserData} from "./userData.js";
+import {Button, Container} from "@material-ui/core";
+import {ActionsContext} from "../../context/ActionsContext";
+import {makeStyles} from "@material-ui/core/styles";
+import {useForm} from "../../hooks/useForm";
 
 const useStyles = makeStyles(theme => ({
-  button: {
-    margin: "2em"
-  }
+    button: {
+        margin: "2em"
+    }
 }));
 
+const initialData = {
+    first_name: "",
+    last_name: "",
+    role: "",
+    phone: "",
+    company: "",
+    company_url: "",
+    about: ""
+};
+
 const UserSettings = () => {
-  const actions = useContext(ActionsContext);
-  const [isEditing, setIsEditing] = useState(false);
-  const { token, currentUser } = useSelector(state => state.user);
-  const styles = useStyles();
-  const [values, handleChange, handleSubmit] = useForm(
-    currentUser.user_metadata,
-    doSubmit
-  );
+    const actions = useContext(ActionsContext);
+    const [isEditing, setIsEditing] = useState(false);
+    const [data, setData] = useState({});
+    const {token, currentUser} = useSelector(state => state.user);
+    const styles = useStyles();
+    const [values, handleChange, handleSubmit] = useForm(
+        currentUser.user_metadata,
+        doSubmit
+    );
 
-  function doSubmit() {
-    console.log(values);
-    actions.user.updateUser(token, values);
-  }
+    useEffect(() => {
+        const checkCurrentUser = async () => {
+            if (!currentUser.user_metadata) {
+                await actions.user.updateUser(token, initialData);
+                setData(initialData);
+            } else {
+                setData(currentUser.user_metadata);
+            }
+        };
+        checkCurrentUser();
+    }, []);
 
-  console.log("current user", currentUser);
-  return (
-    <React.Fragment>
-      <Container maxWidth="lg">
-        {/* <AuthForm /> */}
-        {/* <UserSettingsForm /> */}
-        {!isEditing ? (
-          <UserData />
-        ) : (
-          <UserSettingsForm
-            values={values}
-            handleChange={handleChange}
-            handleSubmit={handleSubmit}
-          />
-        )}
-        {!isEditing && (
-          <Button
-            variant="contained"
-            color="primary"
-            className={styles.button}
-            onClick={() => !isEditing && setIsEditing(true)}
-          >
-            Edit Details
-          </Button>
-        )}
-        {isEditing && (
-          <>
-            <Button
-              variant="contained"
-              color="primary"
-              className={styles.button}
-              onClick={e => {
-                handleSubmit(e);
-                setIsEditing(false);
-              }}
-            >
-              Save Changes
-            </Button>
-            <Button
-              variant="contained"
-              color="secondary"
-              className={styles.button}
-              onClick={() => isEditing && setIsEditing(false)}
-            >
-              Cancel Edit
-            </Button>
-          </>
-        )}
-      </Container>
-    </React.Fragment>
-  );
+    useEffect(() => {
+        setData(currentUser.user_metadata)
+    }, [currentUser.user_metadata]);
+
+    function doSubmit() {
+        setData(values);
+        actions.user.updateUser(token, values);
+        setIsEditing(false)
+    }
+
+    console.log("current user", currentUser);
+    return (
+        <React.Fragment>
+            <Container maxWidth="lg">
+                {/* <AuthForm /> */}
+                {/* <UserSettingsForm /> */}
+                {!isEditing ? (
+                    <UserData data={data} initialData={initialData}/>
+                ) : (
+                    <UserSettingsForm
+                        values={values}
+                        handleChange={handleChange}
+                        handleSubmit={handleSubmit}
+                    />
+                )}
+                {!isEditing && (
+                    <Button
+                        variant="contained"
+                        color="primary"
+                        className={styles.button}
+                        onClick={() => !isEditing && setIsEditing(true)}
+                    >
+                        Edit Details
+                    </Button>
+                )}
+                {isEditing && (
+                    <>
+                        <Button
+                            variant="contained"
+                            color="primary"
+                            className={styles.button}
+                            onClick={handleSubmit}
+                        >
+                            Save Changes
+                        </Button>
+                        <Button
+                            variant="contained"
+                            color="secondary"
+                            className={styles.button}
+                            onClick={() => isEditing && setIsEditing(false)}
+                        >
+                            Cancel Edit
+                        </Button>
+                    </>
+                )}
+            </Container>
+        </React.Fragment>
+    );
 };
 
 export default UserSettings;
